Refactor ParamProvider.Item to separate file
Moved the Item struct from ParamProvider.swift to a new ParamProvider.Item.swift file for better code organization. Also changed recordID property in ParamProvider to private for improved encapsulation.
