+++ Sources/Scout/UI/Event/Param/ParamProvider.swift
@@ -10,7 +10,7 @@ import CloudKit
 class ParamProvider: ObservableObject, Provider {
     @Published var result: ProviderResult<[Item]>?
 
-    let recordID: CKRecord.ID
+    private let recordID: CKRecord.ID
 
     init(recordID: CKRecord.ID) {
         self.recordID = recordID
@@ -21,25 +21,5 @@ class ParamProvider: ObservableObject, Provider {
             .record(for: recordID)["params"]
             .map(Item.fromData)?
             .sorted() ?? []
-    }
-}
-
-extension ParamProvider {
-    struct Item: Identifiable, Comparable, Hashable, CustomStringConvertible {
-        static func < (lhs: Item, rhs: Item) -> Bool {
-            lhs.key < rhs.key
-        }
-
-        let id = UUID()
-        let key: String
-        let value: String
-
-        fileprivate static func fromData(_ data: Data) throws -> [Item] {
-            try JSONDecoder().decode([String: String].self, from: data).map(Item.init)
-        }
-
-        var description: String {
-            "\(key): \(value)"
-        }
     }
 }
